Split Loki assertions and schema into their own files
Move the Loki specific assertions to assertion_loki.go and the Loki schema
types and builders to schema_loki.go, next to the Prometheus equivalents.
Both are pure moves. Also drop the comment above the HTTP service.
